Add Scraped::Request::Strategy class
This is the class that should be subclassed in order to provide a custom
request strategy.
